refactor(chain): #1336 P2 — move write transport into RpcFailoverClient
Move the three per-endpoint WRITE-transport loops out of the base into the
module as D8 explicit (non-folded) methods, preserving every tx nuance 1:1:

- populateAndSignAcrossProviders → populateAndSign (estimate-failover-only-
  if-more-providers, gas-buffer path, single-vs-multi host-only message).
- broadcastSignedTransactionWithFailover → broadcast (isKnownTransactionError
  idempotent short-circuit, exhaustion code-stamp).
- getTransactionReceiptWithFailover → getReceipt (sawNonErrorResponse/null,
  RPC_RECEIPT_LOOKUP_FAILED distinct from RPC_ENDPOINTS_EXHAUSTED).

The ONLY edits inside a moved loop: this.providers/rpcUrls → the live thunks
and this.signPopulatedTransaction → the injected signPopulated callback (D2),
so the module still references no adapter state and holds NO tx-safety state.

Tx-safety orchestration stays on the base and calls thin retained delegators
BY NAME (D3) so the existing spies/stubs keep binding: sendSignedTransaction-
AndWait → this.broadcastSignedTransactionWithFailover; waitForReceiptWith-
Failover → this.getTransactionReceiptWithFailover; populateAndSignV10With-
AllowanceRecovery + sendContractTransaction → this.populateAndSignAcross-
Providers. signPopulatedTransaction, the WAL onBroadcast checkpoint, the
forcedReapprove latch, the set-retry passes, and signerTxSerializer are
untouched. Full dkg-chain suite green (709 pass/1 skip), tx-safety tests
unmodified. Follow-up to #1335.
